test(e2e): move kagent scenarios into their own package
Splits the Kagent suite into test/e2e/kagent with feature-per-file tests
and moves the shared registry setup into an importable StartRegistry
helper. Port-forwards default to an ephemeral local port so test
packages can run concurrently.
