+++ test/e2e/kagent/helpers_test.go
@@ -1,6 +1,6 @@
 //go:build e2e
 
-package e2e
+package kagent
 
 import (
 	"bufio"
@@ -25,6 +25,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
 	registryv1alpha1 "github.com/agentregistry-dev/agentregistry/pkg/api/v1alpha1"
+	e2e "github.com/agentregistry-dev/agentregistry/test/e2e"
 )
 
 const (
@@ -38,255 +39,11 @@ const (
 	kagentRemoteIDKey   = "runtimes.agentregistry.solo.io/kagent/remoteId"
 )
 
-func TestKagentAgent(t *testing.T) {
-	env := newKagentTestEnvironment(t)
-	id := UniqueNameWithPrefix("kagent-agent")
-	secretName := "e2e-" + id + "-secret"
-	runtimeName := "e2e-" + id + "-runtime"
-	modelName := "e2e-" + id + "-model"
-	agentName := "e2e-" + id + "-agent"
-	deploymentName := "e2e-" + id + "-deployment"
-	docs := newKagentScenarioDocs(
-		t,
-		"agent",
-		"Kagent Agent E2E scenario",
-		id,
-	)
-	registerKagentObjectCleanup(t, "agents.kagent.dev", agentName)
-	registerKagentResourceCleanup(t, env, [][]string{
-		{"delete", "deployment", deploymentName},
-		{"delete", "agent", agentName},
-		{"delete", "model", modelName, "--tag", "e2e"},
-		{"delete", "-f", filepath.Join(env.workDir, "runtime.yaml")},
-	})
-
-	docs.Step("Configure the Kagent runtime", "Create a Secret and a Kagent Runtime that reads its bearer token from the Secret API.")
-	env.Apply("runtime.yaml", kagentRuntimeManifest(secretName, runtimeName))
-	docs.Apply(kagentRuntimeManifest(secretName, runtimeName))
-
-	docs.Step("Deploy an Agent", "Create a Model and Agent, then deploy the Agent to Kagent using modelRef.")
-	env.Apply("agent.yaml", kagentAgentManifest(modelName, agentName, ""))
-	env.Apply("agent-deployment.yaml", kagentAgentDeploymentManifest(deploymentName, agentName, runtimeName, modelName))
-	docs.Apply(kagentAgentManifest(modelName, agentName, ""))
-	docs.Apply(kagentAgentDeploymentManifest(deploymentName, agentName, runtimeName, modelName))
-	waitForKagentDeploymentReady(t, env.workDir, env.registryURL, deploymentName)
-	waitForKagentResourceCreated(t, "agents.kagent.dev", agentName)
-	workload := waitForKagentWorkloadCreated(t, agentName)
-	assertContainerEnvironment(t, workload, "MODEL_PROVIDER", "bedrock")
-	assertContainerEnvironment(t, workload, "MODEL_NAME", "anthropic.claude-3-5-sonnet-20241022-v2:0")
-	assertKagentDeploymentRemoteID(t, env, deploymentName, agentName)
-
-	docs.Step("Reapply the Deployment", "Reapply the unchanged Deployment and verify Kagent does not replace the Agent.")
-	metadata := kagentResourceMetadata(t, "agents.kagent.dev", agentName)
-	env.Apply("agent-deployment.yaml", kagentAgentDeploymentManifest(deploymentName, agentName, runtimeName, modelName))
-	docs.Apply(kagentAgentDeploymentManifest(deploymentName, agentName, runtimeName, modelName))
-	assertKagentResourceStable(t, "agents.kagent.dev", agentName, metadata, 10*time.Second)
-
-	docs.Step("Remove the Agent Deployment", "Delete the AgentRegistry Deployment and verify Kagent removes the Agent workload.")
-	docs.Command("arctl delete deployment " + deploymentName)
-	docs.Command("kubectl --context kind-${KIND_CLUSTER_NAME} -n kagent wait --for=delete agent/" + agentName + " --timeout=2m")
-	docs.Command("arctl delete agent " + agentName)
-	docs.Command("arctl delete model " + modelName + " --tag e2e")
-	docs.Command("arctl delete -f - <<EOF\n" + kagentRuntimeManifest(secretName, runtimeName) + "\nEOF")
-	env.DeleteDeployment(deploymentName)
-	waitForKagentResourceDeleted(t, "agents.kagent.dev", agentName)
-	waitForKagentWorkloadDeleted(t, agentName)
-}
-
-func TestKagentMCP(t *testing.T) {
-	env := newKagentTestEnvironment(t)
-	id := UniqueNameWithPrefix("kagent-mcp")
-	secretName := "e2e-" + id + "-secret"
-	runtimeName := "e2e-" + id + "-runtime"
-	remoteName := "e2e-" + id + "-remote"
-	remoteDeploymentName := "e2e-" + id + "-remote-deployment"
-	sourceName := "e2e-" + id + "-source"
-	sourceDeploymentName := "e2e-" + id + "-source-deployment"
-	docs := newKagentScenarioDocs(
-		t,
-		"mcp",
-		"Kagent MCP E2E scenario",
-		id,
-	)
-	registerKagentObjectCleanup(t, "remotemcpservers.kagent.dev", remoteName)
-	registerKagentObjectCleanup(t, "mcpservers.kagent.dev", sourceName)
-	registerKagentResourceCleanup(t, env, [][]string{
-		{"delete", "deployment", remoteDeploymentName},
-		{"delete", "deployment", sourceDeploymentName},
-		{"delete", "mcpserver", remoteName},
-		{"delete", "mcpserver", sourceName},
-		{"delete", "-f", filepath.Join(env.workDir, "runtime.yaml")},
-	})
-
-	docs.Step("Configure the Kagent runtime", "Create the Secret and Runtime shared by the MCPServer deployments.")
-	env.Apply("runtime.yaml", kagentRuntimeManifest(secretName, runtimeName))
-	docs.Apply(kagentRuntimeManifest(secretName, runtimeName))
-
-	docs.Step("Deploy a remote MCPServer", "Deploy a streamable HTTP MCPServer and verify Kagent creates a RemoteMCPServer.")
-	env.Apply("remote-mcp.yaml", kagentRemoteMCPManifest(remoteName))
-	env.Apply("remote-mcp-deployment.yaml", kagentMCPDeploymentManifest(remoteDeploymentName, remoteName, runtimeName))
-	docs.Apply(kagentRemoteMCPManifest(remoteName))
-	docs.Apply(kagentMCPDeploymentManifest(remoteDeploymentName, remoteName, runtimeName))
-	waitForKagentDeploymentReady(t, env.workDir, env.registryURL, remoteDeploymentName)
-	waitForKagentResourceCreated(t, "remotemcpservers.kagent.dev", remoteName)
-	assertKagentDeploymentRemoteID(t, env, remoteDeploymentName, remoteName)
-
-	docs.Step("Deploy a source-backed MCPServer", "Deploy a pinned npm stdio MCPServer and verify its live tools/list response.")
-	env.Apply("source-mcp.yaml", kagentSourceMCPManifest(sourceName))
-	env.Apply("source-mcp-deployment.yaml", kagentMCPDeploymentManifest(sourceDeploymentName, sourceName, runtimeName))
-	docs.Apply(kagentSourceMCPManifest(sourceName))
-	docs.Apply(kagentMCPDeploymentManifest(sourceDeploymentName, sourceName, runtimeName))
-	waitForKagentDeploymentReady(t, env.workDir, env.registryURL, sourceDeploymentName)
-	waitForKagentResourceCreated(t, "mcpservers.kagent.dev", sourceName)
-	waitForKagentWorkloadAvailable(t, sourceName)
-	assert.Contains(t, listKagentMCPTools(t, sourceName), "create_entities")
-	assertKagentDeploymentRemoteID(t, env, sourceDeploymentName, sourceName)
-
-	docs.Step("Remove the MCPServer Deployments", "Delete both Deployments and verify Kagent removes both MCPServer resources.")
-	docs.Command("arctl delete deployment " + remoteDeploymentName)
-	docs.Command("arctl delete deployment " + sourceDeploymentName)
-	docs.Command("kubectl --context kind-${KIND_CLUSTER_NAME} -n kagent wait --for=delete remotemcpserver/" + remoteName + " --timeout=2m")
-	docs.Command("kubectl --context kind-${KIND_CLUSTER_NAME} -n kagent wait --for=delete mcpserver/" + sourceName + " --timeout=2m")
-	docs.Command("arctl delete mcpserver " + remoteName)
-	docs.Command("arctl delete mcpserver " + sourceName)
-	docs.Command("arctl delete -f - <<EOF\n" + kagentRuntimeManifest(secretName, runtimeName) + "\nEOF")
-	env.DeleteDeployment(remoteDeploymentName)
-	env.DeleteDeployment(sourceDeploymentName)
-	waitForKagentResourceDeleted(t, "remotemcpservers.kagent.dev", remoteName)
-	waitForKagentResourceDeleted(t, "mcpservers.kagent.dev", sourceName)
-	waitForKagentWorkloadDeleted(t, sourceName)
-}
-
-func TestKagentDiscovery(t *testing.T) {
-	env := newKagentTestEnvironment(t)
-	id := UniqueNameWithPrefix("kagent-discovery")
-	secretName := "e2e-" + id + "-secret"
-	runtimeName := "e2e-" + id + "-runtime"
-	agentName := "e2e-" + id + "-agent"
-	docs := newKagentScenarioDocs(
-		t,
-		"discovery",
-		"Kagent discovery E2E scenario",
-		id,
-	)
-	cleanupRegistry := registerKagentResourceCleanup(t, env, [][]string{
-		{"delete", "-f", filepath.Join(env.workDir, "runtime.yaml")},
-	})
-	registerKagentObjectCleanup(t, "agents.kagent.dev", agentName)
-
-	docs.Step("Configure the Kagent runtime", "Create the Runtime through which AgentRegistry discovers out-of-band Kagent resources.")
-	env.Apply("runtime.yaml", kagentRuntimeManifest(secretName, runtimeName))
-	docs.Apply(kagentRuntimeManifest(secretName, runtimeName))
-
-	oobManifest := kagentOutOfBandAgentManifest(agentName)
-	docs.Step("Create an out-of-band Kagent Agent", "Apply an Agent directly to Kagent instead of creating an AgentRegistry Deployment.")
-	docs.Command("kubectl --context kind-${KIND_CLUSTER_NAME} apply -f - <<EOF\n" + oobManifest + "\nEOF")
-	applyKagentObject(t, "out-of-band-agent.yaml", oobManifest)
-	discovered := waitForDiscoveredKagentDeployment(t, env.registryURL, runtimeName, agentName, true)
-	require.Equal(t, registryv1alpha1.KindAgent, discovered.Spec.TargetRef.Kind)
-	require.Equal(t, registryv1alpha1.DeploymentOriginDiscovered, discovered.Metadata.Annotations[registryv1alpha1.DeploymentOriginAnnotation])
-
-	docs.Step("Remove the discovered Agent", "Delete the out-of-band Agent and its Runtime, then verify the discovered Deployment is removed.")
-	docs.Command("kubectl --context kind-${KIND_CLUSTER_NAME} -n kagent delete agent " + agentName)
-	docs.Command("arctl delete -f - <<EOF\n" + kagentRuntimeManifest(secretName, runtimeName) + "\nEOF")
-	deleteKagentObject(t, "agents.kagent.dev", agentName)
-	cleanupRegistry()
-	waitForDiscoveredKagentDeployment(t, env.registryURL, runtimeName, agentName, false)
-}
-
-func TestKagentAgentMCP(t *testing.T) {
-	registryURL := RegistryURL(t)
-	workDir := t.TempDir()
-	id := UniqueNameWithPrefix("kagent")
-	runtimeName := "e2e-" + id + "-runtime"
-	secretName := "e2e-" + id + "-secret"
-	modelName := "e2e-" + id + "-model"
-	agentName := "e2e-" + id + "-agent"
-	agentDeploymentName := "e2e-" + id + "-agent-deployment"
-	mcpName := "e2e-" + id + "-mcp"
-	mcpDeploymentName := "e2e-" + id + "-mcp-deployment"
-
-	names := kagentLifecycleNames{
-		Secret:        secretName,
-		Runtime:       runtimeName,
-		Model:         modelName,
-		Agent:         agentName,
-		AgentDeploy:   agentDeploymentName,
-		MCPServer:     mcpName,
-		MCPDeployment: mcpDeploymentName,
-	}
-	manifest := kagentLifecycleManifest(names)
-	docs := newKagentScenarioDocs(
-		t,
-		"agent-mcp",
-		"Kagent Agent with MCP E2E scenario",
-		id,
-	)
-	registerKagentObjectCleanup(t, "agents.kagent.dev", agentName)
-	registerKagentObjectCleanup(t, "mcpservers.kagent.dev", mcpName)
-	docs.Step(
-		"Deploy an Agent with an MCPServer",
-		"Create a secret-backed Kagent Runtime, deploy a source-backed MCPServer, and deploy an Agent that references it without deploymentRefs.",
-	)
-	docs.Apply(manifest)
-	manifestPath := filepath.Join(workDir, "kagent-lifecycle.yaml")
-	require.NoError(t, os.WriteFile(manifestPath, []byte(manifest), 0o600))
-	require.NoError(
-		t,
-		os.WriteFile(
-			filepath.Join(workDir, "runtime.yaml"),
-			[]byte(kagentRuntimeManifest(secretName, runtimeName)),
-			0o600,
-		),
-	)
-	registerKagentRegistryCleanup(t, workDir, registryURL, names)
-
-	RequireSuccess(t, RunArctl(t, workDir, "apply", "-f", manifestPath, "--registry-url", registryURL))
-	waitForKagentDeploymentReady(t, workDir, registryURL, mcpDeploymentName)
-	waitForKagentDeploymentReady(t, workDir, registryURL, agentDeploymentName)
-
-	waitForKagentResourceCreated(t, "agents.kagent.dev", agentName)
-	agentWorkload := waitForKagentWorkloadCreated(t, agentName)
-	assertContainerEnvironment(t, agentWorkload, "HOST", "0.0.0.0")
-	assertContainerEnvironment(t, agentWorkload, "KAGENT_NAMESPACE", kagentNamespace)
-	assertContainerEnvironment(t, agentWorkload, "KAGENT_NAME", agentName)
-	assertContainerEnvironment(t, agentWorkload, "KAGENT_URL", kagentControllerURL)
-	assertContainerEnvironment(t, agentWorkload, "MODEL_PROVIDER", "bedrock")
-	assertContainerEnvironment(t, agentWorkload, "MODEL_NAME", "anthropic.claude-3-5-sonnet-20241022-v2:0")
-
-	waitForKagentResourceCreated(t, "mcpservers.kagent.dev", mcpName)
-	waitForKagentWorkloadAvailable(t, mcpName)
-	assert.Contains(t, listKagentMCPTools(t, mcpName), "create_entities")
-
-	mcpConfig := containerEnvironment(t, agentWorkload, "MCP_SERVERS_CONFIG")
-	require.NotEmpty(t, mcpConfig)
-	assertKagentMCPConfig(t, mcpConfig, mcpDeploymentName, mcpName)
-
-	docs.Step(
-		"Remove the deployments",
-		"Delete both AgentRegistry Deployments and verify that Kagent removes their resources.",
-	)
-	docs.Command("arctl delete deployment " + agentDeploymentName)
-	docs.Command("arctl delete deployment " + mcpDeploymentName)
-	docs.Command("kubectl --context kind-${KIND_CLUSTER_NAME} -n kagent wait --for=delete agent/" + agentName + " --timeout=2m")
-	docs.Command("kubectl --context kind-${KIND_CLUSTER_NAME} -n kagent wait --for=delete mcpserver/" + mcpName + " --timeout=2m")
-	docs.Command("arctl delete agent " + agentName)
-	docs.Command("arctl delete mcpserver " + mcpName)
-	docs.Command("arctl delete model " + modelName + " --tag e2e")
-	docs.Command("arctl delete -f - <<EOF\n" + kagentRuntimeManifest(secretName, runtimeName) + "\nEOF")
-	RequireSuccess(t, RunArctl(t, workDir, "delete", "deployment", agentDeploymentName, "--registry-url", registryURL))
-	RequireSuccess(t, RunArctl(t, workDir, "delete", "deployment", mcpDeploymentName, "--registry-url", registryURL))
-	waitForKagentResourceDeleted(t, "agents.kagent.dev", agentName)
-	waitForKagentResourceDeleted(t, "mcpservers.kagent.dev", mcpName)
-	waitForKagentWorkloadDeleted(t, agentName)
-	waitForKagentWorkloadDeleted(t, mcpName)
-}
-
 func newKagentScenarioDocs(t *testing.T, name, title, id string) *e2eDocRecorder {
 	t.Helper()
 	return newE2EDocRecorder(
 		t,
-		filepath.Join(testDir(), "scenarios", "kagent", name+".md"),
+		filepath.Join(testDir(), "scenarios", name+".md"),
 		title,
 		map[string]string{id: "${E2E_ID}"},
 	)
@@ -303,7 +60,7 @@ func newKagentTestEnvironment(t *testing.T) kagentTestEnvironment {
 	return kagentTestEnvironment{
 		t:           t,
 		workDir:     t.TempDir(),
-		registryURL: RegistryURL(t),
+		registryURL: registryBaseURL(t),
 	}
 }
 
@@ -311,7 +68,7 @@ func (e kagentTestEnvironment) Apply(fileName, manifest string) {
 	e.t.Helper()
 	path := filepath.Join(e.workDir, fileName)
 	require.NoError(e.t, os.WriteFile(path, []byte(manifest), 0o600))
-	RequireSuccess(e.t, RunArctl(e.t, e.workDir, "apply", "-f", path, "--registry-url", e.registryURL))
+	e2e.RequireSuccess(e.t, e2e.RunArctl(e.t, e.workDir, "apply", "-f", path, "--registry-url", e.registryURL))
 }
 
 func (e kagentTestEnvironment) DeleteDeployment(name string) {
@@ -321,7 +78,7 @@ func (e kagentTestEnvironment) DeleteDeployment(name string) {
 
 func (e kagentTestEnvironment) DeleteResource(kind, name string) {
 	e.t.Helper()
-	RequireSuccess(e.t, RunArctl(e.t, e.workDir, "delete", kind, name, "--registry-url", e.registryURL))
+	e2e.RequireSuccess(e.t, e2e.RunArctl(e.t, e.workDir, "delete", kind, name, "--registry-url", e.registryURL))
 }
 
 func registerKagentResourceCleanup(
@@ -335,7 +92,7 @@ func registerKagentResourceCleanup(
 		once.Do(func() {
 			for _, arguments := range resources {
 				arguments = append(arguments, "--registry-url", env.registryURL)
-				checkKagentCleanupResult(t, RunArctl(t, env.workDir, arguments...))
+				checkKagentCleanupResult(t, e2e.RunArctl(t, env.workDir, arguments...))
 			}
 		})
 	}
@@ -488,7 +245,7 @@ func kagentResourceMetadata(t *testing.T, resource, name string) metav1.PartialO
 	t.Helper()
 	cmd := exec.Command(
 		"kubectl",
-		"--context", e2eKubeContext,
+		"--context", e2e.KubeContext,
 		"--namespace", kagentNamespace,
 		"get", resource, name,
 		"--output", "json",
@@ -527,14 +284,14 @@ func assertKagentDeploymentRemoteID(
 	want string,
 ) {
 	t.Helper()
-	result := RunArctl(
+	result := e2e.RunArctl(
 		t,
 		env.workDir,
 		"get", "deployment", deploymentName,
 		"--output", "json",
 		"--registry-url", env.registryURL,
 	)
-	RequireSuccess(t, result)
+	e2e.RequireSuccess(t, result)
 	var deployment registryv1alpha1.Deployment
 	require.NoError(t, json.Unmarshal([]byte(result.Stdout), &deployment))
 	require.Equal(t, want, deployment.Metadata.Annotations[kagentRemoteIDKey])
@@ -544,7 +301,7 @@ func applyKagentObject(t *testing.T, fileName, manifest string) {
 	t.Helper()
 	path := filepath.Join(t.TempDir(), fileName)
 	require.NoError(t, os.WriteFile(path, []byte(manifest), 0o600))
-	output, err := exec.Command("kubectl", "--context", e2eKubeContext, "apply", "-f", path).CombinedOutput()
+	output, err := exec.Command("kubectl", "--context", e2e.KubeContext, "apply", "-f", path).CombinedOutput()
 	require.NoError(t, err, strings.TrimSpace(string(output)))
 }
 
@@ -552,7 +309,7 @@ func deleteKagentObject(t *testing.T, resource, name string) {
 	t.Helper()
 	output, err := exec.Command(
 		"kubectl",
-		"--context", e2eKubeContext,
+		"--context", e2e.KubeContext,
 		"--namespace", kagentNamespace,
 		"delete", resource, name,
 		"--ignore-not-found",
@@ -752,7 +509,7 @@ func waitForKagentResourceCreated(t *testing.T, resource, name string) {
 	require.EventuallyWithT(t, func(c *assert.CollectT) {
 		cmd := exec.Command(
 			"kubectl",
-			"--context", e2eKubeContext,
+			"--context", e2e.KubeContext,
 			"--namespace", kagentNamespace,
 			"get", resource, name,
 			"--ignore-not-found",
@@ -771,7 +528,7 @@ func waitForKagentResourceDeleted(t *testing.T, resource, name string) {
 	require.EventuallyWithT(t, func(c *assert.CollectT) {
 		cmd := exec.Command(
 			"kubectl",
-			"--context", e2eKubeContext,
+			"--context", e2e.KubeContext,
 			"--namespace", kagentNamespace,
 			"get", resource, name,
 			"--ignore-not-found",
@@ -791,7 +548,7 @@ func waitForKagentResourceDeleted(t *testing.T, resource, name string) {
 func waitForKagentDeploymentReady(t *testing.T, workDir, registryURL, name string) {
 	t.Helper()
 	require.EventuallyWithT(t, func(c *assert.CollectT) {
-		result := RunArctl(
+		result := e2e.RunArctl(
 			t,
 			workDir,
 			"get", "deployment", name,
@@ -851,7 +608,7 @@ func waitForKagentWorkloadDeleted(t *testing.T, name string) {
 func getKagentWorkload(name string) (appsv1.Deployment, error) {
 	cmd := exec.Command(
 		"kubectl",
-		"--context", e2eKubeContext,
+		"--context", e2e.KubeContext,
 		"--namespace", kagentNamespace,
 		"get", "deployment", name,
 		"--output", "json",
@@ -946,7 +703,7 @@ func startKagentPortForward(t *testing.T, workload string) int {
 	cmd := exec.CommandContext(
 		ctx,
 		"kubectl",
-		"--context", e2eKubeContext,
+		"--context", e2e.KubeContext,
 		"--namespace", kagentNamespace,
 		"port-forward",
 		"--address", "127.0.0.1",
@@ -1018,7 +775,7 @@ func registerKagentRegistryCleanup(
 			}
 			for _, arguments := range resources {
 				arguments = append(arguments, "--registry-url", registryURL)
-				checkKagentCleanupResult(t, RunArctl(t, workDir, arguments...))
+				checkKagentCleanupResult(t, e2e.RunArctl(t, workDir, arguments...))
 			}
 		})
 	}
@@ -1026,7 +783,7 @@ func registerKagentRegistryCleanup(
 	return cleanup
 }
 
-func checkKagentCleanupResult(t *testing.T, result ArctlResult) {
+func checkKagentCleanupResult(t *testing.T, result e2e.ArctlResult) {
 	t.Helper()
 	if result.ExitCode == 0 {
 		return
