+++ test/e2e/kagent/scenarios/agent-mcp.md
