+++ test/e2e/kagent/scenarios/discovery.md
