+++ test/e2e/kagent/scenarios/mcp.md
